cobalt: make core handle type 32/64bit-neutral
There is no point in using 64bit long Cobalt handles internally on
64bit platforms. However it makes sense to keep them as 32bit values
so that we may use them as atomic operands on either sides of a
64bit(kernel) <-> 32bit(user) ABI.

At this chance, boilerplate/atomic.h is sanitized for disambiguating
regular and long atomic set/get operations.

+++ include/cobalt/uapi/kernel/types.h
@@ -29,7 +29,7 @@ typedef __u64 xntime_t; /* ns */
 
 typedef __s64 xnstime_t;
 
-typedef unsigned long xnhandle_t;
+typedef __u32 xnhandle_t;
 
 #define XN_NO_HANDLE ((xnhandle_t)0)
